Rename <Entity>Id to Scalar<Entity>Id
